// frontend/src/api-client/types.gen.ts
export type BackendInternalModelFormSummary = {
    createdAt?: string;
    description?: string;
    id?: string;
    publicUrl?: string;
    published?: boolean;
    responsesCount?: number;
    title?: string;
};